Title: 10개 입력받기, Time: 131ms, Memory: 10MB, Status: Passed - Codetree

+++ 250913/10개 입력받기/receive-10-inputs.java
@@ -17,7 +17,7 @@ public class Main {
             sum+= arr[i];
             cnt++;
         }  
-        avg=sum/cnt;
+        avg=(float)sum/cnt;
         System.out.printf("%d %.1f",sum,avg);
     }
 }
